+++ app/helpers/airbrake_helper.rb
@@ -3,18 +3,13 @@ module AirbrakeHelper
   def format_table(data)
     lines = []
     data.each do |key, value|
-      next unless value.is_a?(String)
-      lines << "|@#{key}@|#{value.strip.blank? ? value : "@#{value}@"}|"
-    end
-    lines.join("\n")
-  end
+      next if value.blank?
 
-  # Wiki markup for logs
-  def format_log(data)
-    lines = []
-    data.each do |log|
-      next unless log.is_a?(Hash)
-      lines << "[#{log[:time].strftime('%F %T')}] #{log[:line]}"
+      if value.is_a?(String)
+        lines << "|@#{key}@|@#{value}@|"
+      elsif value.is_a?(Hash)
+        lines << "|@#{key}@|@#{value.map { |k, v| "#{k}: #{v}"}.join(',')}@|"
+      end
     end
     lines.join("\n")
   end
@@ -31,28 +26,28 @@ module AirbrakeHelper
     repository = repository_for_backtrace_element(element)
 
     if repository.blank?
-      if element[:number].blank?
-        markup = "@#{element[:file]}@"
+      if element.line.blank?
+        markup = "@#{element.file}@"
       else
-        markup = "@#{element[:file]}:#{element[:number]}@"
+        markup = "@#{element.file}:#{element.line}@"
       end
     else
       filename = filename_for_backtrace_element(element)
 
       if repository.identifier.blank?
-        markup = "source:\"#{filename}#L#{element[:number]}\""
+        markup = "source:\"#{filename}#L#{element.line}\""
       else
-        markup = "source:\"#{repository.identifier}|#{filename}#L#{element[:number]}\""
+        markup = "source:\"#{repository.identifier}|#{filename}#L#{element.line}\""
       end
     end
 
-    markup + " in ??<notextile>#{element[:method]}</notextile>??"
+    markup + " in ??<notextile>#{element.function}</notextile>??"
   end
 
   private
 
   def repository_for_backtrace_element(element)
-    return nil unless element[:file].start_with?('[PROJECT_ROOT]')
+    return nil unless element.file.start_with?('[PROJECT_ROOT]')
 
     filename = filename_for_backtrace_element(element)
 
@@ -62,18 +57,18 @@ module AirbrakeHelper
   def repositories_for_backtrace
     return @_bactrace_repositories unless @_bactrace_repositories.nil?
 
-    if @request.repository.present?
-      @_bactrace_repositories = [@request.repository]
+    if @repository.present?
+      @_bactrace_repositories = [@repository]
     else
-      @_bactrace_repositories = @request.project.repositories.to_a
+      @_bactrace_repositories = @project.repositories.to_a
     end
 
     @_bactrace_repositories
   end
 
   def filename_for_backtrace_element(element)
-    return nil if  element[:file].blank?
+    return nil if  element.file.blank?
 
-    element[:file][14..-1]
+    element.file[14..-1]
   end
 end